requirements.txt
pandas
h5py
opencv-python
matplotlib
numpy
tqdm
torch
torchvision
monai
argparse
pyyaml